spl_token_group_interface

Module instruction

Instruction types

Structs

InitializeGroup

Instruction data for initializing a new Group

InitializeMember

Instruction data for initializing a new Member of a Group

UpdateGroupAuthority

Instruction data for updating the authority of a Group

UpdateGroupMaxSize

Instruction data for updating the max size of a Group

Enums

TokenGroupInstruction

All instructions that must be implemented in the SPL Token Group Interface

Functions

initialize_group

Creates a InitializeGroup instruction

initialize_member

Creates a InitializeMember instruction

update_group_authority

Creates a UpdateGroupAuthority instruction

update_group_max_size

Creates a UpdateGroupMaxSize instruction