spl_token_group_interface::instruction

Enum TokenGroupInstruction

pub enum TokenGroupInstruction {
    InitializeGroup(InitializeGroup),
    UpdateGroupMaxSize(UpdateGroupMaxSize),
    UpdateGroupAuthority(UpdateGroupAuthority),
    InitializeMember(InitializeMember),
}

All instructions that must be implemented in the SPL Token Group Interface

Variants

InitializeGroup(InitializeGroup)

Initialize a new Group

Assumes one has already initialized a mint for the group.

Accounts expected by this instruction:

[w] Group

[] Mint

[s] Mint authority

UpdateGroupMaxSize(UpdateGroupMaxSize)

Update the max size of a Group

Accounts expected by this instruction:

[w] Group

[s] Update authority

UpdateGroupAuthority(UpdateGroupAuthority)

Update the authority of a Group

Accounts expected by this instruction:

[w] Group

[s] Current update authority

InitializeMember(InitializeMember)

Initialize a new Member of a Group

Assumes the Group has already been initialized, as well as the mint for the member.

Accounts expected by this instruction:

[w] Member

[] Member mint

[s] Member mint authority

[w] Group

[s] Group update authority

Implementations

impl TokenGroupInstruction

pub fn unpack(input: &[u8]) -> Result<Self, ProgramError>

Unpacks a byte buffer into a TokenGroupInstruction

pub fn pack(&self) -> Vec<u8>

Packs a TokenGroupInstruction into a byte buffer.